feat: updated delete user handler

+++ src/handlers/user/delete.js
@@ -1,7 +1,15 @@
+import mongoose from 'mongoose';
 import { userModel } from '../../models/user.js';
+import { accepted, badRequest } from '../../utils/responses.js';
 
 export const deleteUserHandler = async (req, res) => {
   const { userId } = req.params;
+
+  if (!mongoose.isValidObjectId(userId)) {
+    return badRequest(res, 'Invalid userId');
+  }
+
   await userModel.findByIdAndDelete(userId);
-  res.status(202).send('Request accepted!');
+
+  return accepted();
 };
